Création Liste des prêts de l'utilisateur Récupération properties du microservice Ajout de la fonctionalité de renouvellement

+++ library/src/main/java/com/library/beans/mbooks/lending/LendingBean.java
@@ -28,7 +28,7 @@ public  @Data class LendingBean {
     @DateTimeFormat(pattern="yyyy-MM-dd HH:mm:ss")
     private Date returnDate;
 
-    private Long renewal;
+    private Integer renewal;
 
     @NotNull(message = "Le choix du livre est obligatoire")
     private BookBean book;
